requirements.txt
amqp==5.0.1
asgiref==3.2.10
bcrypt==3.1.4
billiard==3.6.3.0
boto3==1.15.0
botocore==1.18.0
CacheControl==0.12.6
celery==5.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
deprecation==2.1.0
Django==3.1.2
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.11.0
drf-writable-nested==0.6.1
drf-yasg==1.17.1
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==0.9.6
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.0
namegenerator==1.0.6
numpy==1.19.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pandas==1.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
smartystreets-python-sdk==4.6.1
sqlparse==0.3.1
squareup==6.1.0.20200722
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
